BIG MODS: refactoring and using pane now
