+++ shapash/report/core.py
@@ -7,10 +7,11 @@ import logging
 from pathlib import Path
 
 import pandas as pd
+import panel as pn
 
-from shapash.report.smart_report.blocks import ReportBlockMixin
-from shapash.report.smart_report.layout import build_html_page, render_sections, resolve_logo_src
-from shapash.report.smart_report.validation import load_report_config, render_block_error_html
+from shapash.report.blocks import ReportBlockMixin
+from shapash.report.panel_support import apply_report_css
+from shapash.report.validation import load_report_config, render_block_error
 
 logger = logging.getLogger(__name__)
 
@@ -50,52 +51,76 @@ def create_block_runtime(
 
 
 def generate_report(runtime, config_file: str, output_file: str) -> None:
-    """Render a report to an HTML file driven by a YAML config."""
+    """Render a Panel report to an HTML file driven by a YAML config."""
     cfg_path = Path(config_file).resolve()
     cfg = load_report_config(cfg_path)
     print(f"Loading config → {cfg_path}")
 
     runtime.render_block = lambda block_cfg: render_block(runtime, block_cfg)
-    rendered_blocks, sidebar_html = render_sections(runtime, cfg["sections"])
+    rendered_blocks = [render_block(runtime, block_cfg) for block_cfg in cfg["sections"]]
 
     out_path = Path(output_file).resolve()
     out_path.parent.mkdir(parents=True, exist_ok=True)
-    body_html = "\n".join(rendered_blocks)
-    logo_src = resolve_logo_src(out_path.parent)
-    out_path.write_text(build_html_page(body=body_html, sidebar_html=sidebar_html, logo_src=logo_src), encoding="utf-8")
+
+    apply_report_css()
+    report_layout = pn.Column(
+        *[block for block in rendered_blocks if block is not None],
+        css_classes=["main-report"],
+        sizing_mode="stretch_width",
+    )
+    report_layout.save(str(out_path), embed=True, resources="cdn")
     logger.info("Report saved → %s", output_file)
 
 
-def render_block(runtime, block_cfg: dict) -> str:
+def render_block(runtime, block_cfg: dict):
     """Dispatch one YAML block entry to the matching block_* method."""
     block_type = block_cfg.get("type", "")
     params = block_cfg.get("params", {})
 
     if block_type == "group":
-        return "".join(render_block(runtime, child_cfg) for child_cfg in block_cfg.get("blocks", []))
+        children = [render_block(runtime, child_cfg) for child_cfg in block_cfg.get("blocks", [])]
+        children = [child for child in children if child is not None]
+        group_title = params.get("title", "")
+        if group_title:
+            return pn.Column(
+                pn.pane.Markdown(f"## {group_title}", css_classes=["group-title"]),
+                *children,
+                sizing_mode="stretch_width",
+            )
+        return pn.Column(*children, sizing_mode="stretch_width")
 
     method = getattr(runtime, f"block_{block_type}", None)
     if method is None:
         if block_type == "custom":
             return _render_custom(runtime, block_cfg)
         logger.warning("Unknown block type '%s' — skipped.", block_type)
-        return ""
+        return None
 
     try:
-        return method(**params)
+        result = method(**params)
+        if isinstance(result, pn.viewable.Viewable):
+            return result
+        if isinstance(result, str):
+            return pn.pane.Markdown(result)
+        return pn.panel(result)
     except Exception as exc:
         logger.error("Block '%s' raised: %s", block_type, exc)
-        return render_block_error_html(block_type, exc)
+        return render_block_error(block_type, exc)
 
 
-def _render_custom(runtime, block_cfg: dict) -> str:
+def _render_custom(runtime, block_cfg: dict):
     """Call an arbitrary importable function."""
     func_path = block_cfg.get("function", "")
     params = block_cfg.get("params", {})
     try:
         mod_path, fn_name = func_path.rsplit(".", 1)
         fn = getattr(importlib.import_module(mod_path), fn_name)
-        return fn(runtime, **params)
+        result = fn(runtime, **params)
+        if isinstance(result, pn.viewable.Viewable):
+            return result
+        if isinstance(result, str):
+            return pn.pane.Markdown(result)
+        return pn.panel(result)
     except Exception as exc:
         logger.error("Custom block '%s' raised: %s", func_path, exc)
-        return render_block_error_html(func_path, exc)
+        return render_block_error(func_path, exc)
