--- shapash/report/smart_report/report_styles.css
@@ -1,396 +1,0 @@
-/* Theme tokens */
-:root {
-    --shapash-gold: #f4c000;
-    --text-main: #343736;
-    --text-light: #777;
-    --sidebar-width: 240px;
-}
-
-/* Reset */
-*,
-*::before,
-*::after {
-    box-sizing: border-box;
-    margin: 0;
-    padding: 0;
-}
-
-html {
-    scroll-behavior: smooth;
-}
-
-/* Page shell */
-body,
-.report-shell {
-    background: #fdfdfd;
-    color: var(--text-main);
-    font-family: "Helvetica Neue", Arial, sans-serif;
-}
-
-.report-shell {
-    display: flex;
-    min-height: 100vh;
-    width: 100%;
-}
-
-/* Sidebar */
-.sidebar {
-    width: var(--sidebar-width);
-    background: #fff;
-    border-right: 1px solid #eee;
-    position: fixed;
-    height: 100vh;
-    padding: 30px 20px;
-    overflow-y: auto;
-    box-shadow: inset -8px 0 12px -12px rgba(0, 0, 0, 0.45);
-}
-
-.sidebar-brand {
-    margin-bottom: 40px;
-    display: flex;
-    align-items: center;
-    gap: 10px;
-    color: var(--shapash-gold);
-    font-size: 18px;
-    font-weight: bold;
-}
-
-.sidebar-brand-logo {
-    display: block;
-    width: 34px;
-    height: 34px;
-    flex: 0 0 auto;
-}
-
-.sidebar-brand-text {
-    line-height: 1.2;
-}
-
-/* Navigation */
-.nav-item {
-    color: var(--text-light);
-    padding: 10px 0;
-    display: block;
-    text-decoration: none;
-    font-size: 13px;
-    transition: 0.2s;
-}
-
-.nav-item:hover {
-    color: var(--text-main);
-}
-
-.nav-item.active {
-    color: #551a8b;
-    font-weight: bold;
-    border-left: 3px solid #551a8b;
-    padding-left: 10px;
-}
-
-.nav-group {
-    margin-bottom: 2px;
-}
-
-.nav-group-title {
-    color: var(--text-main);
-    font-size: 13px;
-    font-weight: 500;
-    padding: 10px 0;
-    display: block;
-    text-decoration: none;
-    transition: 0.2s;
-}
-
-.nav-group-title:hover {
-    color: #551a8b;
-}
-
-.nav-group-title.active {
-    color: #551a8b;
-    font-weight: bold;
-    border-left: 3px solid #551a8b;
-    padding-left: 10px;
-}
-
-.nav-children {
-    display: block;
-}
-
-.nav-child {
-    color: var(--text-light);
-    padding: 6px 0 6px 16px;
-    display: block;
-    text-decoration: none;
-    font-size: 12px;
-    transition: 0.2s;
-}
-
-.nav-child:hover {
-    color: var(--text-main);
-}
-
-.nav-child.active {
-    color: #551a8b;
-    font-weight: 600;
-    border-left: 3px solid #551a8b;
-    padding-left: 13px;
-}
-
-.nav-level-2 {
-    padding-left: 28px;
-}
-
-.nav-level-2.active {
-    padding-left: 25px;
-}
-
-.nav-level-3 {
-    padding-left: 40px;
-}
-
-.nav-level-3.active {
-    padding-left: 37px;
-}
-
-.nav-level-4 {
-    padding-left: 52px;
-}
-
-.nav-level-4.active {
-    padding-left: 49px;
-}
-
-/* Main content */
-.container {
-    margin-left: var(--sidebar-width);
-    width: calc(100% - var(--sidebar-width));
-    padding: 60px 80px 60vh;
-    max-width: 1400px;
-}
-
-.main-header {
-    text-align: center;
-    margin-bottom: 60px;
-}
-
-.main-header h1 {
-    font-size: 2.4rem;
-    font-weight: 500;
-    color: #343736;
-    margin-bottom: 20px;
-}
-
-.section-title {
-    font-size: 1.6rem;
-    color: #343736;
-    margin: 40px 0 20px;
-    font-weight: 700;
-}
-
-.section-block {
-    margin-bottom: 30px;
-}
-
-.content-block {
-    margin-bottom: 30px;
-    line-height: 1.6;
-    font-size: 14px;
-}
-
-/* Callouts and separators */
-.shapash-callout {
-    border-left: 4px solid var(--shapash-gold);
-    background: #fff;
-    padding: 15px 25px;
-    margin: 30px 0;
-    color: #343736;
-    line-height: 1.6;
-    font-size: 15px;
-}
-
-.shapash-divider {
-    border-bottom: 1px solid #eee;
-    margin: 50px 0;
-}
-
-/* Generic key/value and dataframe tables */
-.kv-table,
-table.dataframe {
-    width: 100%;
-    border-collapse: separate;
-    border-spacing: 0;
-    margin: 12px 0 24px;
-    background: #fff;
-    border: 1px solid #ececec;
-    border-radius: 12px;
-    overflow: hidden;
-    box-shadow: 0 8px 24px rgba(0, 0, 0, 0.04);
-}
-
-.project-info-table {
-    table-layout: fixed;
-}
-
-.project-info-table .kv-key {
-    width: 42%;
-    min-width: 280px;
-    padding-left: 18px;
-    padding-right: 10px;
-}
-
-.project-info-table .kv-key-label,
-.project-info-table .kv-key-sep {
-    overflow-wrap: normal;
-    word-break: keep-all;
-    hyphens: none;
-}
-
-.project-info-table .kv-val {
-    padding-left: 10px;
-    padding-right: 18px;
-}
-
-.analysis-side-by-side {
-    display: flex;
-    gap: 20px;
-    align-items: flex-start;
-    flex-wrap: wrap;
-}
-
-.analysis-side-table {
-    flex: 1 1 360px;
-    min-width: 320px;
-}
-
-.analysis-side-plot {
-    flex: 1 1 480px;
-    min-width: 320px;
-}
-
-.analysis-side-table .kv-table {
-    margin-top: 0;
-}
-
-/* Plot wrappers */
-.scroll-section {
-    scroll-margin-top: 40px;
-}
-
-.panel-plot {
-    width: 100%;
-    overflow-x: auto;
-}
-
-.panel-plot .bk-root,
-.panel-plot .plotly-graph-div {
-    width: 100%;
-    max-width: 100%;
-}
-
-/* Desktop-only responsive adjustments */
-@media (max-width: 1600px) and (min-width: 1281px) {
-    .container {
-        padding: 52px 60px 50vh;
-    }
-}
-
-@media (max-width: 1366px) and (min-width: 1024px) {
-    .container {
-        padding: 44px 44px 44vh;
-    }
-
-    .main-header {
-        margin-bottom: 48px;
-    }
-
-    .main-header h1 {
-        font-size: 2.2rem;
-    }
-
-    .model-analysis-tables {
-        gap: 16px;
-    }
-
-    .model-analysis-table-col {
-        flex: 0 1 calc(50% - 8px);
-    }
-
-    .analysis-side-by-side {
-        gap: 16px;
-    }
-}
-
-@media (max-width: 1200px) and (min-width: 1024px) {
-    :root {
-        --sidebar-width: 220px;
-    }
-
-    .sidebar {
-        width: var(--sidebar-width);
-        padding: 24px 16px;
-    }
-
-    .container {
-        margin-left: var(--sidebar-width);
-        width: calc(100% - var(--sidebar-width));
-        padding: 34px 30px 38vh;
-    }
-
-    .section-title {
-        font-size: 1.45rem;
-    }
-
-    .analysis-side-table {
-        min-width: 300px;
-    }
-
-    .analysis-side-plot {
-        min-width: 300px;
-    }
-
-    .kv-table,
-    table.dataframe {
-        display: block;
-        overflow-x: auto;
-        -webkit-overflow-scrolling: touch;
-    }
-}
-
-@media (max-width: 1100px) and (min-width: 1024px) {
-    :root {
-        --sidebar-width: 200px;
-    }
-
-    .sidebar-brand {
-        font-size: 16px;
-        margin-bottom: 28px;
-    }
-
-    .nav-item,
-    .nav-group-title {
-        font-size: 12px;
-    }
-
-    .model-analysis-table-col {
-        flex: 1 1 100%;
-        min-width: 0;
-    }
-
-    .analysis-side-table,
-    .analysis-side-plot {
-        flex: 1 1 100%;
-        min-width: 0;
-    }
-
-    .univariate-select {
-        min-width: 0;
-        width: 100%;
-    }
-
-    .content-block,
-    .shapash-callout,
-    .kv-val {
-        overflow-wrap: anywhere;
-        word-break: break-word;
-    }
-}
